+++ yz-mall-sys/yz-mall-sys-service/src/main/java/com/yz/mall/sys/entity/SysRole.java
@@ -45,7 +45,7 @@ public class SysRole extends Model<SysRole> {
      * 数据是否有效：0数据有效
      */
     @TableLogic
-    private Integer invalid;
+    private Long invalid;
 
     /**
      * 角色编码
