+++ yz-unqid/yz-unqid-core/src/main/java/com/yz/unqid/entity/SysUnqid.java
@@ -41,7 +41,7 @@ public class SysUnqid implements Serializable {
      * 数据是否有效：0数据有效
      */
     @TableLogic
-    private Integer invalid;
+    private Long invalid;
 
     /**
      * 流水号的前缀
